Lazy Load Guarantee Component and Set State on Load
The Guarantee component is being lazy loaded because there isn't a
guaranteee (pun intended) that it will ever be clicked.  No need to
fetch data if never loaded. This saves on bandwith and loading time.

There is currently one issue that I can think of.  Currently, when the
Guarantee is called a get request is made. This request will be made on
each display of the Guarantee component.  The Guaranteee state should
only need to fetch data once.  If the data was fetched, store the data
in state and pass it down. Construct a check, only fetch if state is not
set, and display the stored state.